+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - styles for toolbar
 - styles for overallInfo
 - styles for day budget
+- styles for savings
 
 ## [3.3.0] - 2022-01-06
 ### Added